Fixing imports

+++ models/loot.go
@@ -2,7 +2,7 @@ package models
 
 // The location and value of loot pulled from the database
 type Loot struct {
-	Latitude  float
-	Longitude float
+	Latitude  float64
+	Longitude float64
 	Value     int
 }
